restructure benchmarks
